Split replies out to a different file to reduce noise
