+++ meeseeks/meeseeks.go
@@ -5,7 +5,6 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/pcarranza/meeseeks-box/auth"
-	"github.com/pcarranza/meeseeks-box/command"
 	"github.com/pcarranza/meeseeks-box/config"
 	"github.com/pcarranza/meeseeks-box/meeseeks/commands"
 	"github.com/pcarranza/meeseeks-box/meeseeks/message"
@@ -86,83 +85,4 @@ func (m Meeseeks) Process(msg message.Message) {
 		req.Username)
 	m.replyWithSuccess(j.Request, cmd, out)
 	j.Finish(jobs.SuccessStatus)
-}
-
-func (m Meeseeks) replyWithInvalidMessage(msg message.Message, err error) {
-	content, err := m.templates.Build().RenderFailure(msg.GetUsernameID(), err.Error(), "")
-	if err != nil {
-		log.Fatalf("could not render failure template: %s", err)
-	}
-
-	if err = m.client.Reply(content, m.config.Colors.Error, msg.GetChannelID()); err != nil {
-		log.Errorf("Failed to reply: %s", err)
-	}
-}
-
-func (m Meeseeks) replyWithUnknownCommand(req request.Request) {
-	log.Debugf("Could not find command '%s' in the command registry", req.Command)
-
-	msg, err := m.templates.Build().RenderUnknownCommand(req.UsernameID, req.Command)
-	if err != nil {
-		log.Fatalf("could not render unknown command template: %s", err)
-	}
-
-	if err = m.client.Reply(msg, m.config.Colors.Error, req.ChannelID); err != nil {
-		log.Errorf("Failed to reply: %s", err)
-	}
-}
-
-func (m Meeseeks) replyWithHandshake(req request.Request, cmd command.Command) {
-	if !cmd.HasHandshake() {
-		return
-	}
-	msg, err := m.buildTemplatesFor(cmd).RenderHandshake(req.UsernameID)
-	if err != nil {
-		log.Fatalf("could not render unknown command template: %s", err)
-	}
-
-	if err = m.client.Reply(msg, m.config.Colors.Info, req.ChannelID); err != nil {
-		log.Errorf("Failed to reply: %s", err)
-	}
-}
-
-func (m Meeseeks) replyWithUnauthorizedCommand(req request.Request, cmd command.Command) {
-	log.Debugf("User %s is not allowed to run command '%s' on channel '%s'", req.Username,
-		req.Command, req.Channel)
-
-	msg, err := m.buildTemplatesFor(cmd).RenderUnauthorizedCommand(req.UsernameID, req.Command)
-	if err != nil {
-		log.Fatalf("could not render unathorized command template %s", err)
-	}
-
-	if err = m.client.Reply(msg, m.config.Colors.Error, req.ChannelID); err != nil {
-		log.Errorf("Failed to reply: %s", err)
-	}
-}
-
-func (m Meeseeks) replyWithCommandFailed(req request.Request, cmd command.Command, err error, out string) {
-	msg, err := m.buildTemplatesFor(cmd).RenderFailure(req.UsernameID, err.Error(), out)
-	if err != nil {
-		log.Fatalf("could not render failure template %s", err)
-	}
-
-	if err = m.client.Reply(msg, m.config.Colors.Error, req.ChannelID); err != nil {
-		log.Errorf("Failed to reply: %s", err)
-	}
-}
-
-func (m Meeseeks) replyWithSuccess(req request.Request, cmd command.Command, out string) {
-	msg, err := m.buildTemplatesFor(cmd).RenderSuccess(req.UsernameID, out)
-
-	if err != nil {
-		log.Fatalf("could not render success template %s", err)
-	}
-
-	if err = m.client.Reply(msg, m.config.Colors.Success, req.ChannelID); err != nil {
-		log.Errorf("Failed to reply: %s", err)
-	}
-}
-
-func (m Meeseeks) buildTemplatesFor(cmd command.Command) template.Templates {
-	return m.templates.Clone().WithTemplates(cmd.Templates()).Build()
 }
